requirements.txt
# torch>=1.10.0
# torchvision
segmentation-models-pytorch
timm
matplotlib
pandas
opencv-python
tqdm